// frontend/src/app/core/services/api-client/model/pacienteRes.ts
/**
 * Telemedicina API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {Persona} from './persona';


export interface PacienteRes {
    id?: number;
    persona?: Persona;
}